refactor: split node renderer viewport root helpers
